CLN: Move _aggregate and _aggregate_multiple_funcs to core.aggregation
